\name{countCharOccurrences}
\alias{countCharOccurrences}
\title{count char occurrences}
\usage{
countCharOccurrences(char, s)
}
\arguments{
  \item{char}{the character to find}
  \item{s}{the string to find the char in}
}

\description{
Function for counting occurences of a char in string. Takes 2 arguments, the first the character to count, the second is the string that contains this character. It returns an integer.
}
\examples{
countCharOccurrences("s", "this is some text that has the letter s in it.")
}
